backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic>=2.0.0
python-multipart==0.0.6

# Core OpenRouter API integration
openai>=1.40.0
python-dotenv>=1.0.0

# HTTP Clients
aiohttp>=3.8.0
httpx>=0.26.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Excel export functionality
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Async operations and rate limiting
asyncio-throttle>=1.0.0
backoff>=2.2.0

# Utilities
rich>=13.0.0
python-dateutil>=2.8.0
jsonschema>=4.17.0

# Analysis and ML
textstat>=0.7.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional web utilities
pytz>=2023.3